refactor: modularize firebase setup
